requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
psycopg2-binary==2.9.9